+++ packages/create-sitecore-jss/src/templates/nextjs-sxa/src/pages/404.tsx
@@ -40,7 +40,7 @@ export const getStaticProps: GetStaticProps = async (context) => {
   });
   let resultErrorPages: ErrorPages | null = null;
 
-  if (!process.env.DISABLE_SSG_FETCH) {
+  if (process.env.DISABLE_SSG_FETCH?.toLowerCase() !== 'true') {
     try {
       resultErrorPages = await errorPagesService.fetchErrorPages();
     } catch (error) {