+++ packages/sitecore-jss-nextjs/src/editing/editing-data-service.test.ts
@@ -2,7 +2,7 @@
 /* eslint-disable no-unused-expressions */
 import { expect, use } from 'chai';
 import chaiAsPromised from 'chai-as-promised';
-import { AxiosDataFetcher } from '@sitecore-jss/sitecore-jss';
+import { NativeDataFetcher } from '@sitecore-jss/sitecore-jss';
 import { EditingData } from './editing-data';
 import { EditingDataCache } from './editing-data-cache';
 import {
@@ -18,7 +18,7 @@ use(sinonChai);
 use(chaiAsPromised);
 
 const mockFetcher = (data?: unknown) => {
-  const fetcher = {} as AxiosDataFetcher;
+  const fetcher = {} as NativeDataFetcher;
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   fetcher.get = spy<any>(() => {
     return Promise.resolve({ data });