+++ packages/sitecore-jss-nextjs/src/editing/editing-render-middleware.test.ts
@@ -3,7 +3,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { expect, use } from 'chai';
 import { NextApiRequest, NextApiResponse } from 'next';
-import { AxiosDataFetcher } from '@sitecore-jss/sitecore-jss';
+import { NativeDataFetcher } from '@sitecore-jss/sitecore-jss';
 import { EditingDataService, EditingPreviewData } from './editing-data-service';
 import {
   EDITING_ALLOWED_ORIGINS,
@@ -80,7 +80,7 @@ const mockResponse = () => {
 };
 
 const mockFetcher = (html?: string) => {
-  const fetcher = {} as AxiosDataFetcher;
+  const fetcher = {} as NativeDataFetcher;
   fetcher.get = spy<any>(() => {
     return Promise.resolve({ data: html ?? '' });
   });
@@ -457,7 +457,7 @@ describe('EditingRenderMiddleware', () => {
       query[QUERY_PARAM_EDITING_SECRET] = secret;
       const previewData = { key: 'key1234' } as EditingPreviewData;
 
-      const fetcher = {} as AxiosDataFetcher;
+      const fetcher = {} as NativeDataFetcher;
       fetcher.get = spy<any>(() => {
         return Promise.reject({ response: { data: html, status: 404 } });
       });
@@ -500,7 +500,7 @@ describe('EditingRenderMiddleware', () => {
       query[QUERY_PARAM_EDITING_SECRET] = secret;
       const previewData = { key: 'key1234' } as EditingPreviewData;
 
-      const fetcher = {} as AxiosDataFetcher;
+      const fetcher = {} as NativeDataFetcher;
       fetcher.get = spy<any>(() => {
         return Promise.reject({ response: { data: html, status: 500 } });
       });