+++ packages/sitecore-jss-react/src/index.ts
@@ -5,6 +5,11 @@ export {
   CacheClient,
   CacheOptions,
   MemoryCacheClient,
+  HttpDataFetcher,
+  NativeDataFetcher,
+  NativeDataFetcherError,
+  NativeDataFetcherResponse,
+  NativeDataFetcherConfig,
 } from '@sitecore-jss/sitecore-jss';
 export { EnhancedOmit } from '@sitecore-jss/sitecore-jss/utils';
 export {